+++ src/handling.rs
@@ -279,6 +279,9 @@ impl MessageHandler {
             .iter()
             //filter out my ID to avoid loopback
             .filter(|&n| &n.id != self.my_header.binary_id().as_binary())
+            // Limit the number of peers accepted in a single Nodes message to a
+            // reasonable bound
+            .take(K_K * 2)
             .filter(|&n| {
                 let h = self.my_header.binary_id().calculate_distance(&n.id);
                 match h {