+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move to stable MSRV 1.85
+- Move to rust edition 2024
 
 ## [0.7.0] - 2024-10-21